test: refactor fetchFormEnums
