+++ bciers/apps/administration/tests/components/facilities/FacilitiesPage.test.tsx
@@ -1,6 +1,6 @@
+import { getOperation } from "@bciers/testConfig/mocks";
 import { render, screen } from "@testing-library/react";
 import FacilitiesPage from "apps/administration/app/components/facilities/FacilitiesPage";
-import { getOperation } from "../operations/mocks";
 
 const searchParams = { operations_title: "Operation 2" };
 // mocking the child component until this issue is fixed: https://github.com/testing-library/react-testing-library/issues/1209#issuecomment-1673372612
