+++ bciers/apps/administration/tests/components/userOperators/UserOperator.test.tsx
@@ -1,8 +1,11 @@
 import { render, screen } from "@testing-library/react";
-import { auth, useSession } from "@bciers/testConfig/mocks";
+import {
+  auth,
+  getBusinessStructures,
+  useSession,
+} from "@bciers/testConfig/mocks";
 import { getUserOperatorFormData } from "./mocks";
 import UserOperator from "apps/administration/app/components/userOperators/UserOperator";
-import { getBusinessStructures } from "../operations/mocks";
 
 describe("UserOperator component", () => {
   beforeEach(async () => {
