+++ bciers/apps/administration/vitest.config.mts
@@ -27,8 +27,6 @@ export default defineConfig({
     },
     setupFiles: [
       "../../libs/testConfig/src/global.tsx",
-      "./tests/components/facilities/mocks.ts",
-      "./tests/components/operations/mocks.ts",
       "./tests/components/operators/mocks.ts",
       "./tests/components/contacts/mocks.ts",
     ],
