+++ bciers/apps/registration/tests/components/operations/registration/FacilityInformationPage.test.tsx
@@ -2,7 +2,7 @@ import { render, screen } from "@testing-library/react";
 import FacilityInformationPage from "apps/registration/app/components/operations/registration/FacilityInformationPage";
 import { allOperationRegistrationSteps } from "@/registration/app/components/operations/registration/enums";
 import {
-  actionHandler,
+  fetchFacilitiesPageData,
   useSearchParams,
   useSession,
 } from "@bciers/testConfig/mocks";
@@ -29,7 +29,7 @@ describe("the FacilityInformationPage component", () => {
     vi.clearAllMocks();
   });
   it("should render the FacilityInformationPage component", async () => {
-    actionHandler.mockReturnValueOnce({
+    fetchFacilitiesPageData.mockReturnValueOnce({
       items: [],
       count: 0,
     });
@@ -54,9 +54,9 @@ describe("the FacilityInformationPage component", () => {
   });
 
   it("should render the single facility operation form with name and type pre-populated", async () => {
-    actionHandler.mockReturnValueOnce({
-      items: [],
-      count: 0,
+    fetchFacilitiesPageData.mockReturnValueOnce({
+      rows: [],
+      row_count: 0,
     });
     const { container } = render(
       await FacilityInformationPage({
@@ -85,8 +85,8 @@ describe("the FacilityInformationPage component", () => {
   });
 
   it("should render the datagrid with facility information", async () => {
-    actionHandler.mockReturnValueOnce({
-      items: [
+    fetchFacilitiesPageData.mockReturnValueOnce({
+      rows: [
         {
           id: "1",
           facility__name: "Test Facility 1",
@@ -100,7 +100,7 @@ describe("the FacilityInformationPage component", () => {
           status: "Active",
         },
       ],
-      count: 1,
+      row_count: 1,
     });
     render(
       await FacilityInformationPage({
@@ -121,7 +121,7 @@ describe("the FacilityInformationPage component", () => {
   });
 
   it("should render the empty datagrid when there is no facility data", async () => {
-    actionHandler.mockReturnValueOnce({
+    fetchFacilitiesPageData.mockReturnValueOnce({
       items: [],
       count: 0,
     });
