+++ bciers/apps/registration/tests/components/operations/registration/OperationInformationPage.test.tsx
@@ -3,7 +3,8 @@ import OperationInformationPage from "apps/registration/app/components/operation
 import { useSession } from "@bciers/testConfig/mocks";
 import { allOperationRegistrationSteps } from "@/registration/app/components/operations/registration/enums";
 
-import { fetchFormEnums } from "../OperationRegistrationPage.test";
+import fetchFormEnums from "@bciers/testConfig/helpers/fetchFormEnums";
+import { Apps } from "@bciers/utils/src/enums";
 
 useSession.mockReturnValue({
   data: {
@@ -18,7 +19,7 @@ describe("the OperationInformationPage component", () => {
     vi.clearAllMocks();
   });
   it("should render the OperationInformationPage component", async () => {
-    fetchFormEnums();
+    fetchFormEnums(Apps.REGISTRATION);
 
     render(
       await OperationInformationPage({
